+++ src/Extractor/ClassMethodToCallbackExtractor.php
@@ -7,10 +7,13 @@ namespace AutoMapper\Extractor;
 use AutoMapper\Exception\InvalidArgumentException;
 use AutoMapper\Exception\LogicException;
 use AutoMapper\Exception\RuntimeException;
+use PhpParser\Node;
 use PhpParser\Node\Arg;
 use PhpParser\Node\Expr;
 use PhpParser\Node\Param;
 use PhpParser\Node\Stmt;
+use PhpParser\NodeTraverser;
+use PhpParser\NodeVisitor\NameResolver;
 use PhpParser\Parser;
 use PhpParser\ParserFactory;
 
@@ -48,9 +51,12 @@ final readonly class ClassMethodToCallbackExtractor
         }
 
         $statements = $this->parser->parse($fileContents);
+
         if (null === $statements) {
             throw new RuntimeException("Couldn't parse file \"{$fileName}\" for class \"{$class}\".");
         }
+
+        $statements = $this->resolveFullyQualifiedClassNames($statements);
 
         $namespaceStatement = self::findUnique(Stmt\Namespace_::class, $statements, $fileName);
         /** @var Stmt\Class_ $classStatement */
@@ -80,18 +86,18 @@ final readonly class ClassMethodToCallbackExtractor
     }
 
     /**
-     * @template T of Stmt
+     * @template T of Node
      *
      * @param class-string<T> $searchedStatementClass
-     * @param Stmt[]          $statements
+     * @param Node[]          $statements
      *
      * @return T
      */
-    private static function findUnique(string $searchedStatementClass, array $statements, string $fileName): Stmt
+    private static function findUnique(string $searchedStatementClass, array $statements, string $fileName): Node
     {
         $foundStatements = array_filter(
             $statements,
-            static fn (Stmt $statement): bool => $statement instanceof $searchedStatementClass,
+            static fn (Node $statement): bool => $statement instanceof $searchedStatementClass,
         );
 
         if (\count($foundStatements) > 1) {
@@ -99,5 +105,20 @@ final readonly class ClassMethodToCallbackExtractor
         }
 
         return array_values($foundStatements)[0] ?? throw new InvalidArgumentException("No \"{$searchedStatementClass}\" found in file \"{$fileName}\".");
+    }
+
+    /**
+     * Transform all statements with imported class names, into FQCNs.
+     *
+     * @param Node[] $statements
+     *
+     * @return Node[]
+     */
+    private function resolveFullyQualifiedClassNames(array $statements): array
+    {
+        $nodeTraverser = new NodeTraverser();
+        $nodeTraverser->addVisitor(new NameResolver());
+
+        return $nodeTraverser->traverse($statements);
     }
 }