+++ tests/Fixtures/Transformer/CustomTransformer/FromTargetCustomModelTransformer.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace AutoMapper\Tests\Fixtures\Transformer\CustomTransformer;
 
+use AutoMapper\Tests\Fixtures\AddressDTO;
 use AutoMapper\Transformer\CustomTransformer\CustomModelTransformerInterface;
 use Symfony\Component\PropertyInfo\Type;
 
@@ -19,7 +20,7 @@ final readonly class FromTargetCustomModelTransformer implements CustomModelTran
      */
     public function transform(object|array $source): mixed
     {
-        $addressDTO = new \AutoMapper\Tests\Fixtures\AddressDTO();
+        $addressDTO = new AddressDTO();
         $addressDTO->city = "{$source['city']} from custom model transformer";
 
         return $addressDTO;
@@ -45,7 +46,7 @@ final readonly class FromTargetCustomModelTransformer implements CustomModelTran
     private function targetIsAddressDTO(array $targetTypes): bool
     {
         foreach ($targetTypes as $targetType) {
-            if ($targetType->getClassName() === \AutoMapper\Tests\Fixtures\AddressDTO::class) {
+            if ($targetType->getClassName() === AddressDTO::class) {
                 return true;
             }
         }